change preview js to preview tsx
